Update Webpage
Updated some things on the template, after the template updated the pages based off of the template to help reduce work.  Added styles 2c-grid and 3c-grid they should be self explanatory, if not they are respectively a two column grid layout and a three column layout. Other than that there are some minor fixes and QOL (Quality of Life) changes.